Make componentsMap private like standardsMap
