+++ lib/workspace.go
@@ -7,10 +7,6 @@ import (
 	"path/filepath"
 	"sync"
 	"github.com/opencontrol/compliance-masonry/lib/components/versions/base"
-	"github.com/opencontrol/compliance-masonry/lib/components/versions"
-	"github.com/opencontrol/compliance-masonry/lib/components"
-	"github.com/opencontrol/compliance-masonry/tools/constants"
-	"github.com/opencontrol/compliance-masonry/tools/fs"
 	"github.com/codegangsta/cli"
 	"github.com/opencontrol/compliance-masonry/lib/standards"
 	"github.com/opencontrol/compliance-masonry/lib/certifications"
@@ -32,7 +28,7 @@ type Workspace interface {
 // LocalWorkspace struct combines components, standards, and a certification data
 // For more information on the opencontrol schema visit: https://github.com/opencontrol/schemas
 type LocalWorkspace struct {
-	componentsMap  *components.Components
+	components     *componentsMap
 	standards      *standardsMap
 	justifications *Justifications
 	certification  certifications.Certification
@@ -48,8 +44,8 @@ func getKey(filePath string) string {
 func NewWorkspace() Workspace {
 	return &LocalWorkspace{
 		justifications: NewJustifications(),
-		componentsMap:  components.NewComponents(),
-		standards:      NewStandards(),
+		components:  newComponents(),
+		standards:      newStandards(),
 	}
 }
 
@@ -127,40 +123,12 @@ func (ws *LocalWorkspace) LoadStandards(standardsDir string) []error {
 }
 
 
-// LoadComponent imports components into a Component struct and adds it to the
-// Components map.
-func (ws *LocalWorkspace) LoadComponent(componentDir string) error {
-	// Get file system assistance.
-	fs := fs.OSUtil{}
-	// Read the component file.
-	fileName := filepath.Join(componentDir, "component.yaml")
-	componentData, err := fs.OpenAndReadFile(fileName)
-	if err != nil {
-		return errors.New(constants.ErrComponentFileDNE)
-	}
-	// Parse the component.
-	var component base.Component
-	component, err = versions.ParseComponent(componentData,fileName)
-	if err != nil {
-		return err
-	}
-	// Ensure we have a key for the component.
-	if component.GetKey() == "" {
-		component.SetKey(getKey(componentDir))
-	}
-	// If the component is new, make sure we load the justifications as well.
-	if ws.componentsMap.CompareAndAdd(component) {
-		ws.justifications.LoadMappings(component)
-	}
-	return nil
-}
-
 func (ws *LocalWorkspace) GetComponent(component string) base.Component {
-	return ws.componentsMap.Get(component)
+	return ws.components.get(component)
 }
 
 func (ws *LocalWorkspace) GetAllComponents() []base.Component {
-	return ws.componentsMap.GetAll()
+	return ws.components.getAll()
 }
 
 func (ws *LocalWorkspace) GetCertification() certifications.Certification {
